+++ cmd/build_test.go
@@ -41,12 +41,12 @@ svc2
 			assert.Equal(t, expOut, stdout)
 		})
 
-		t.Run("build with private registry 403 with service def", func(t *testing.T) {
+		t.Run("build with private registry 403 with module def", func(t *testing.T) {
 			os.Setenv("MUSS_TEST_REGISTRY_ERROR", "403")
 			defer os.Unsetenv("MUSS_TEST_REGISTRY_ERROR")
 
 			cfg := newTestConfig(t, map[string]interface{}{
-				"service_definitions": []map[string]interface{}{
+				"module_definitions": []map[string]interface{}{
 					map[string]interface{}{
 						"name": "app",
 						"configs": map[string]interface{}{
@@ -75,7 +75,7 @@ svc2
 			defer os.Unsetenv("MUSS_TEST_REGISTRY_ERROR")
 
 			cfg := newTestConfig(t, map[string]interface{}{
-				"service_definitions": []map[string]interface{}{
+				"module_definitions": []map[string]interface{}{
 					map[string]interface{}{
 						"name": "app",
 						"configs": map[string]interface{}{