+++ cmd/pull_test.go
@@ -45,12 +45,12 @@ svc
 			assert.Equal(t, expOut, stdout)
 		})
 
-		t.Run("pull with private registry 403 with service def", func(t *testing.T) {
+		t.Run("pull with private registry 403 with module def", func(t *testing.T) {
 			os.Setenv("MUSS_TEST_REGISTRY_ERROR", "403")
 			defer os.Unsetenv("MUSS_TEST_REGISTRY_ERROR")
 
 			cfg := newTestConfig(t, map[string]interface{}{
-				"service_definitions": []map[string]interface{}{
+				"module_definitions": []map[string]interface{}{
 					map[string]interface{}{
 						"name": "app",
 						"configs": map[string]interface{}{